+++ src/Core/UI/Common/ItemExplorerUI.cs
@@ -369,9 +369,9 @@ namespace StardustSandbox.Core.UI.Common
 
             int index = 0;
 
-            for (byte col = 0; col < UIConstants.ITEM_EXPLORER_ITEMS_PER_COLUMN; col++)
+            for (int col = 0; col < UIConstants.ITEM_EXPLORER_ITEMS_PER_COLUMN; col++)
             {
-                for (byte row = 0; row < UIConstants.ITEM_EXPLORER_ITEMS_PER_ROW; row++)
+                for (int row = 0; row < UIConstants.ITEM_EXPLORER_ITEMS_PER_ROW; row++)
                 {
                     SlotInfo slot = new(
                         new()
